Prepare campaign for C23_WMDE_Mobile_DE_10 Based on VAR of test 09
Ticket: https://phabricator.wikimedia.org/T351696
