+++ campaign_info.toml
@@ -33,9 +33,9 @@ resolution = ["800x600", "1024x768", "1280x960", "1600x1200", "1920x1200", "2560
 [mobile]
 name = "Mobile"
 icon = "mobile"
-campaign = "C23_WMDE_Mobile_DE_09"
-description = "based on CTRL 08, VAR has 3 buttons in soft close"
-campaign_tracking = "mob-de-09-ba-231115"
+campaign = "C23_WMDE_Mobile_DE_10"
+description = "based on VAR 09, Use the word hi, informal text"
+campaign_tracking = "mob-de-10-ba-231123"
 preview_link = "/mobile/wiki/Wikipedia:Hauptseite?devbanner={{banner}}&banner=B22_WMDE_local_prototype&useskin=minerva"
 preview_url = 'https://de.m.wikipedia.org/wiki/Wikipedia:Hauptseite?banner={{banner}}&useskin=minerva&devMode'
 wrapper_template = "wikipedia_org"
@@ -44,13 +44,13 @@ use_of_funds_source = "MediaWiki:WMDE_Fundraising/UseOfFunds_2023_DE"
 # Banners of the campaign, key after "banners" can be anything
 [mobile.banners.ctrl]
 filename = "./banners/mobile/banner_ctrl.ts"
-pagename = "B23_WMDE_Mobile_DE_09_ctrl"
-tracking = "org-mob09-231115-ctrl"
+pagename = "B23_WMDE_Mobile_DE_10_ctrl"
+tracking = "org-mob10-231123-ctrl"
 
 [mobile.banners.var]
 filename = "./banners/mobile/banner_var.ts"
-pagename = "B23_WMDE_Mobile_DE_09_var"
-tracking = "org-mob09-231115-var"
+pagename = "B23_WMDE_Mobile_DE_10_var"
+tracking = "org-mob10-231123-var"
 
 [mobile.test_matrix]
 device = [ 'samsung_s10', 'iphone_xs_max', 'iphone_5s', 'iphone_se', "iphone_8", "iphone_12_mini", "iphone_7_plus", "iphone_11_pro_max"]
